"Fixed" the issue of the conditions of !executeAction not being true
This was caused because the environment would perform clearPercepts()
before the agent had a chance to calculate its action. I fixed it so
that the beliefs that are set as conditions for the !executeAction plan
now get stored with [source(self)] to ensure that they always exist,
even between steps. Of course, this isn't really a fix because stuff
like visibleEdge beliefs still get deleted by clearPercepts(), which
will cause the agent to not perform the best action it could. But at
least the agent should now do something, instead of nothing.
Moved +position trigger to storeBelief.asl
Changed !executeAction to +requestAction as the trigger for performing
an action - there was no reason to waste a Jason cycle on switching
between the two.
Added current Energy to +requestAction print output.