Fix: inspect failing because of failed_resources
It seems that the cost for inspecting increases by 1 if the enemy to be
inspected makes a move away from the inspector in the same turn that we
try to inspect him. This doesn't happen very often, but it's an easy
enough fix, so I'm adding 1 to the minimum energy required to inspect.